+++ lib/src/extensions/color_extensions.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+// Color Extensions
 extension Hex on Color {
   /// return hex String
   String toHex({bool leadingHashSign = true, bool includeAlpha = false}) =>