+++ lib/src/extensions/double_extensions.dart
@@ -2,6 +2,7 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 
+// Double Extensions
 extension DoubleExtensions on double? {
   /// Validate given double is not null and returns given value if null.
   double validate({double value = 0.0}) => this ?? value;