+++ lib/src/extensions/duration_extensions.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+// Duration Extensions
 extension GetDurationUtils on Duration {
   ///  await Duration(seconds: 1).delay();
   Future<void> get delay => Future.delayed(this);