+++ lib/src/extensions/num_extensions.dart
@@ -1,5 +1,6 @@
 import 'package:nb_utils/nb_utils.dart';
 
+// num Extensions
 extension NumExt on num? {
   /// Validate given double is not null and returns given value if null.
   num validate({num value = 0}) => this ?? value;