+++ lib/src/models/walkthrough_model.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+/// Model class for walkthrough
 class WalkThroughModelClass {
   String? title;
   String? subTitle;