+++ lib/src/utils/size_config.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+/// Model class for Size Configurations
 class SizeConfig {
   static MediaQueryData? _mediaQueryData;
   static double? screenWidth;