+++ lib/src/widgets/app_text_field.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:nb_utils/nb_utils.dart';
 
+/// Enum for Text Field
 enum TextFieldType {
   EMAIL,
   PASSWORD,